0-putchar.c

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	long char a;
+	unsigned char a;
 
 	a = '_putchar';
